# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bringing group travel plans to life",
        "title": "Toggo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/files/confirm": {
            "post": {
                "description": "Verifies the file exists in S3 and marks the upload as confirmed in DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Confirm upload",
                "operationId": "confirmUpload",
                "parameters": [
                    {
                        "description": "Confirm upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConfirmUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConfirmUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/trips/{tripID}/pitches": {
            "get": {
                "description": "Returns pitches for the trip with cursor-based pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pitches"
                ],
                "summary": "Get all pitches for a trip",
                "operationId": "listPitches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max items per page (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque cursor from previous response next_cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PitchCursorPageResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new pitch for the trip and returns a presigned URL to upload the audio file.\nOptionally supply up to 5 confirmed image IDs (image_ids) in the request. The response includes images with presigned medium_url for each image.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pitches"
                ],
                "summary": "Create a pitch",
                "operationId": "createPitch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create pitch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePitchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePitchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/trips/{tripID}/pitches/{pitchID}": {
            "get": {
                "description": "Returns a single pitch with presigned URLs for the audio file and associated images (medium_url for each image)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pitches"
                ],
                "summary": "Get a pitch by ID",
                "operationId": "getPitch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pitch ID",
                        "name": "pitchID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PitchAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a pitch by ID",
                "tags": [
                    "pitches"
                ],
                "summary": "Delete a pitch",
                "operationId": "deletePitch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pitch ID",
                        "name": "pitchID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates pitch metadata (title, description, duration).\nWhen image_ids is omitted from the request, existing image associations remain unchanged.\nWhen image_ids is an empty array, all image associations are removed.\nWhen image_ids is a non-empty array, the provided IDs are added to existing associations (additive, not replacement).\nThe response includes images with presigned medium_url for each image.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pitches"
                ],
                "summary": "Update a pitch",
                "operationId": "updatePitch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pitch ID",
                        "name": "pitchID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update pitch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePitchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PitchAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/trips/{tripID}/pitches/{pitchID}/confirm-upload": {
            "post": {
                "description": "Verifies that the audio file was successfully uploaded to S3 and notifies trip members.\nMust be called by the pitch creator after the presigned PUT upload completes.",
                "tags": [
                    "pitches"
                ],
                "summary": "Confirm pitch audio upload",
                "operationId": "confirmPitchUpload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "tripID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pitch ID",
                        "name": "pitchID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.APIError": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.ConfirmUploadRequest": {
            "type": "object",
            "required": [
                "imageId"
            ],
            "properties": {
                "imageId": {
                    "type": "string"
                },
                "size": {
                    "description": "Optional: if nil, confirm all sizes",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ImageSize"
                        }
                    ]
                }
            }
        },
        "models.ConfirmUploadResponse": {
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "integer"
                },
                "imageId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.CreatePitchRequest": {
            "type": "object",
            "required": [
                "content_length",
                "content_type",
                "title"
            ],
            "properties": {
                "content_length": {
                    "type": "integer",
                    "minimum": 1
                },
                "content_type": {
                    "type": "string",
                    "minLength": 1
                },
                "description": {
                    "type": "string"
                },
                "image_ids": {
                    "type": "array",
                    "maxItems": 5,
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "models.CreatePitchResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "pitch": {
                    "$ref": "#/definitions/models.PitchAPIResponse"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
...
        "models.ImageSize": {
            "type": "string",
            "enum": [
                "large",
                "medium",
                "small"
            ],
            "x-enum-varnames": [
                "ImageSizeLarge",
                "ImageSizeMedium",
                "ImageSizeSmall"
            ]
        },
...
        "models.PitchAPIResponse": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PitchImageInfo"
                    }
                },
                "title": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.PitchCursorPageResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PitchAPIResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "models.PitchImageInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "medium_url": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdatePitchRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer",
                    "minimum": 0
                },
                "image_ids": {
                    "description": "ImageIDs, when non-nil, fully replaces the pitch's image associations.\nPass an empty slice to remove all images; omit the field to leave images unchanged.",
                    "type": "array",
                    "maxItems": 5,
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
    }
}